refactor: changed file name MainViewModel in HomeViewModel
