+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teamc/KoinModule.kt
@@ -10,6 +10,7 @@ import com.ivanmorgillo.corsoandroid.teamc.detail.RecipeDetailViewModel
 import com.ivanmorgillo.corsoandroid.teamc.favourite.FavouriteViewModel
 import com.ivanmorgillo.corsoandroid.teamc.firebase.Tracking
 import com.ivanmorgillo.corsoandroid.teamc.firebase.TrackingImpl
+import com.ivanmorgillo.corsoandroid.teamc.home.MainViewModel
 import com.ivanmorgillo.corsoandroid.teamc.login.LoginViewModel
 import com.ivanmorgillo.corsoandroid.teamc.search.RecipeSearchViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
