+++ app/src/main/java/com/ivanmorgillo/corsoandroid/teamc/home/HomeFragment.kt
@@ -8,20 +8,16 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.ivanmorgillo.corsoandroid.teamc.MainActivity
-import com.ivanmorgillo.corsoandroid.teamc.MainScreenAction
-import com.ivanmorgillo.corsoandroid.teamc.MainScreenAction.NavigateToDetail
-import com.ivanmorgillo.corsoandroid.teamc.MainScreenEvent
-import com.ivanmorgillo.corsoandroid.teamc.MainScreenEvent.OnFavouriteClicked
-import com.ivanmorgillo.corsoandroid.teamc.MainScreenEvent.OnRecipeClick
-import com.ivanmorgillo.corsoandroid.teamc.MainScreenEvent.OnRefreshClick
-import com.ivanmorgillo.corsoandroid.teamc.MainScreenStates
-import com.ivanmorgillo.corsoandroid.teamc.MainViewModel
 import com.ivanmorgillo.corsoandroid.teamc.R
 import com.ivanmorgillo.corsoandroid.teamc.databinding.FragmentHomeBinding
 import com.ivanmorgillo.corsoandroid.teamc.exhaustive
 import com.ivanmorgillo.corsoandroid.teamc.gone
 import com.ivanmorgillo.corsoandroid.teamc.home.HomeFragmentDirections.Companion.actionHomeFragmentToDetailFragment
 import com.ivanmorgillo.corsoandroid.teamc.home.HomeFragmentDirections.Companion.actionHomeFragmentToSearchFragment
+import com.ivanmorgillo.corsoandroid.teamc.home.MainScreenAction.NavigateToDetail
+import com.ivanmorgillo.corsoandroid.teamc.home.MainScreenEvent.OnFavouriteClicked
+import com.ivanmorgillo.corsoandroid.teamc.home.MainScreenEvent.OnRecipeClick
+import com.ivanmorgillo.corsoandroid.teamc.home.MainScreenEvent.OnRefreshClick
 import com.ivanmorgillo.corsoandroid.teamc.utils.bindings.viewBinding
 import com.ivanmorgillo.corsoandroid.teamc.visible
 import org.koin.androidx.viewmodel.ext.android.viewModel
